moved Logo.tsx to assets directory. For future reference - move each icon to assets directory
